+++ CHANGELOG.md
@@ -21,6 +21,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 * `bList`:
   * Renamed `interface/Option` -> `interface/Item`
   * Renamed `valueProp` -> `itemsProp`
+  * Changed a type of `hideLabels` from a prop to modifier
 
 * `gHint`
   * Renamed: